+++ src/cli.rs
@@ -113,11 +113,13 @@ where
     pub re_filter: Option<Regex>,
 
     /// Tag expression to filter scenarios by.
+    ///
+    /// Note: Tags from Feature, Rule and Scenario are merged together on
+    /// filtering, so be careful about conflicting tags on different levels.
     #[structopt(
         short = "t",
         long = "tags",
         name = "tagexpr",
-        visible_alias = "scenario-tags",
         conflicts_with = "regex"
     )]
     pub tags_filter: Option<TagOperation>,
@@ -173,9 +175,7 @@ This struct is especially useful, when implementing custom [`Writer`] wrapping
 another one:
 ```rust
 # use async_trait::async_trait;
-# use cucumber::{
-#     cli, event, parser, ArbitraryWriter, Event, FailureWriter, World, Writer,
-# };
+# use cucumber::{cli, event, parser, writer, Event, World, Writer};
 # use structopt::StructOpt;
 #
 struct CustomWriter<Wr>(Wr);
@@ -208,11 +208,11 @@ where
 // Useful blanket impls:
 
 #[async_trait(?Send)]
-impl<'val, W, Wr, Val> ArbitraryWriter<'val, W, Val> for CustomWriter<Wr>
+impl<'val, W, Wr, Val> writer::Arbitrary<'val, W, Val> for CustomWriter<Wr>
 where
     W: World,
     Self: Writer<W>,
-    Wr: ArbitraryWriter<'val, W, Val>,
+    Wr: writer::Arbitrary<'val, W, Val>,
     Val: 'val,
 {
     async fn write(&mut self, val: Val)
@@ -223,11 +223,11 @@ where
     }
 }
 
-impl<W, Wr> FailureWriter<W> for CustomWriter<Wr>
+impl<W, Wr> writer::Failure<W> for CustomWriter<Wr>
 where
     W: World,
     Self: Writer<W>,
-    Wr: FailureWriter<W>,
+    Wr: writer::Failure<W>,
 {
     fn failed_steps(&self) -> usize {
         self.0.failed_steps()
@@ -241,6 +241,12 @@ where
         self.0.hook_errors()
     }
 }
+
+impl<Wr: writer::Normalized> writer::Normalized for CustomWriter<Wr> {}
+
+impl<Wr: writer::NonTransforming> writer::NonTransforming
+    for CustomWriter<Wr>
+{}
 ```
 
 [`Writer`]: crate::Writer